backend/requirements.txt
-i https://pypi.org/simple
asgiref==3.7.2 ; python_version >= '3.7'
babel==2.12.1 ; python_version >= '3.7'
boto3==1.26.149
botocore==1.29.149 ; python_version >= '3.7'
dj-database-url==2.0.0
django==4.2.2
django-cockroachdb==4.2
django-cors-headers==4.0.0
django-extensions==3.2.3
django-money==3.1.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
jmespath==1.0.1 ; python_version >= '3.7'
pillow==9.5.0
psycopg[binary]==3.1.9
psycopg-binary==3.1.9
py-moneyed==2.0 ; python_version >= '3.6'
pyjwt==2.7.0 ; python_version >= '3.7'
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2023.3
s3transfer==0.6.1 ; python_version >= '3.7'
setuptools==67.8.0 ; python_version >= '3.7'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.4 ; python_version >= '3.5'
typing-extensions==4.6.3 ; python_version >= '3.7'
urllib3==1.26.16 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
